pyproject.toml
[tool.poetry]
name = "gavno-kod"
version = "0.1.0"
description = ""
authors = ["0leksandrr <tsybulskyy.sasha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pygithub = "^2.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
